Add the program for polynomial solver using tinygrad
The solver used the given data_train.csv to solve for the coefficients of the fourth degree polynomial model. To verify, the model is compared to the data in the data_test.csv

The function is modeled using tinygrad and its SGD method. The parameters for its optimization such as the learning rate,  the loss function and epoch were already chosen such that it gives the best performance. The learning rate = 0.0001 gives the most accurate coefficients. The mean absolute error was used as the basis because it gives the smallest error on the test data, and there are little to no outliers in the dataset. Also, past epoch = 1500, there is not much improvement in the error anymore.

When run, the program will display the polynomial degree and coefficients, and its performance (plot and loss) comparison with data_train.csv and the data_test.csv